Check for existing L2 geth data directory

+++ app/data/config/fixturenet-optimism/run-op-geth.sh
@@ -4,34 +4,39 @@ if [ -n "$CERC_SCRIPT_DEBUG" ]; then
   set -x
 fi
 
-mkdir datadir
-
-echo "pwd" > datadir/password
-
 # TODO: Add in container build or use other tool
 echo "Installing jq"
 apk update && apk add jq
 
-# Get SEQUENCER KEY from keys.json
-SEQUENCER_KEY=$(jq -r '.Sequencer.privateKey' /l2-accounts/keys.json | tr -d '"')
-echo $SEQUENCER_KEY > datadir/block-signer-key
+# Initialize op-geth if datadir/geth not found
+if [ -f /op-node/jwt.txt ] && [ -d datadir/geth ]; then
+  echo "Found existing datadir, skipping initialization"
+else
+  echo "Initializing op-geth"
 
-geth account import --datadir=datadir --password=datadir/password datadir/block-signer-key
+  mkdir datadir
+  echo "pwd" > datadir/password
 
-while [ ! -f "/op-node/jwt.txt" ]
-do
-  echo "Config files not created. Checking after 5 seconds."
-  sleep 5
-done
+  # Get SEQUENCER KEY from keys.json
+  SEQUENCER_KEY=$(jq -r '.Sequencer.privateKey' /l2-accounts/keys.json | tr -d '"')
+  echo $SEQUENCER_KEY > datadir/block-signer-key
 
-echo "Config files created by op-node, proceeding with script..."
+  geth account import --datadir=datadir --password=datadir/password datadir/block-signer-key
 
-cp /op-node/genesis.json ./
-geth init --datadir=datadir genesis.json
+  while [ ! -f "/op-node/jwt.txt" ]
+  do
+    echo "Config files not created. Checking after 5 seconds."
+    sleep 5
+  done
+
+  echo "Config files created by op-node, proceeding with the initialization..."
+
+  geth init --datadir=datadir /op-node/genesis.json
+  echo "Node Initialized"
+fi
 
 SEQUENCER_ADDRESS=$(jq -r '.Sequencer.address' /l2-accounts/keys.json | tr -d '"')
 echo "SEQUENCER_ADDRESS: ${SEQUENCER_ADDRESS}"
-cp /op-node/jwt.txt ./
 
 # Run op-geth
 geth \
@@ -54,7 +59,7 @@ geth \
   --authrpc.vhosts="*" \
   --authrpc.addr=0.0.0.0 \
   --authrpc.port=8551 \
-  --authrpc.jwtsecret=./jwt.txt \
+  --authrpc.jwtsecret=/op-node/jwt.txt \
   --rollup.disabletxpoolgossip=true \
   --password=./datadir/password \
   --allow-insecure-unlock \
